requirements.txt
numpy==1.13.1
pandas==0.20.3
pkg-resources==0.0.0
py==1.4.34
PyQt5==5.9
pytest==3.1.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
sip==4.19.3
six==1.10.0
XlsxWriter==0.9.8
